Enable ODBC Build On Linux
- Enable Linux ODBC build with unicode support
- Added `docker-compose` for Flight SQL ODBC
- Register ODBC after build
- Replaced `boost::lexicographical_compare` with `std::lexicographical_compare`
- Fixed conversion bugs in `SetAttributeSQLWCHAR`
- Convert from std::string to std::u16string directly without involving wide string (wstring)
* Add patch for Windows build

+++ cpp/src/arrow/flight/sql/odbc/odbc_impl/accessors/binary_array_accessor_test.cc
@@ -42,7 +42,7 @@ TEST(BinaryArrayAccessor, Test_CDataType_BINARY_Basic) {
             accessor.GetColumnarData(&binding, 0, values.size(), value_offset, false,
                                      diagnostics, nullptr));
 
-  for (int i = 0; i < values.size(); ++i) {
+  for (size_t i = 0; i < values.size(); ++i) {
     ASSERT_EQ(values[i].length(), str_len_buffer[i]);
     // Beware that CDataType_BINARY values are not null terminated.
     // It's safe to create a std::string from this data because we know it's
